# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WasaText
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text or GIFs and
    easily stay in touch through your private chats or group discussions.
  version: 0.0.1
...
paths:
...
  /profiles/{user}/username:
    parameters:
      - $ref: "#/components/parameters/user"
    put:
      tags: [User]
      operationId: setMyUserName
      security:
        - bearerAuth: []
      summary: set user name of the utent
      description: |
        Taking a name from the requestBody, replace the
        username actually used from the utent with the new one
      requestBody:
        content:
          application/json:
            schema:
              description: New name, to replace the current user name
              type: object
              properties:
                new_name: { $ref: "#/components/schemas/Name" }
        required: true
      responses:
        "200":
          description: The user name was set successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /profiles/{user}/photo:
    parameters:
      - $ref: "#/components/parameters/user"
    put:
      tags: [User]
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      summary: set profile image
      description: |
        Set the profile image of the user
      requestBody:
        content:
          application/json:
            schema:
              description: Photo to upload in the profile picture
              type: object
              properties:
                picture: { $ref: "#/components/schemas/Photo" }
        required: true
      responses:
        "200":
          description: The photo was uploaded succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Completed" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # Tag Group
...
  /profiles/{user}/conversations/{dest_user_id}/messages:
    parameters:
      - $ref: "#/components/parameters/user"
      - $ref: "#/components/parameters/conv"
    post:
      parameters:
        - name: messaggio
          description: messaggio da rispondere
          in: query
          required: false
          schema: { $ref: "#/components/schemas/Identification" }
      tags: [Message]
      operationId: sendMessage
      security:
        - bearerAuth: []
      summary: Send message in a conversation
      description: Send the message in a specified conversation, defined in the path
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SentMessage" }
      responses:
        "201":
          description: Message sended succesfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  parameters:
    user:
      name: user
      in: path
      required: true
      description: The id of the utent who want execute the opreation defined
      schema: { $ref: "#/components/schemas/Identification" }
...
    conv:
      name: dest_user_id
      in: path
      required: true
      description: ID dell'utente o del gruppo con cui aprire la conversazione
      schema: { $ref: "#/components/schemas/Identification" }
...
  schemas:
...
    Completed:
      title: Completed
      description: The operation was successful
      type: string
      pattern: "^.*?$"
      example: "Action successfully completed"
      minLength: 2
      maxLength: 30
    # Elements
    Identification:
      title: ID - identification
      description: General identification number (used for example for users, groups or conversation)
      type: integer
      example: 100
      minimum: 1
    Photo:
      title: Photo
      description: Photo of a user profile ora a group
      type: string
      format: byte
      minLength: 0
      maxLength: 9999
    Name:
      title: Name
      description: Name of a user or a group
      type: string
      pattern: ^[a-z0-9]{3,13}$
      example: "admin123"
      minLength: 3
      maxLength: 13
    Timestamp:
      title: Timestamp
      description: The timestamp of the post.
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
    Profile:
      title: Profile
      description: User profile information
      type: object
      properties:
        username: { $ref: "#/components/schemas/Name" }
        userimg: { $ref: "#/components/schemas/Photo" }
        userid: { $ref: "#/components/schemas/Identification" }
      required:
        - username
        - userid
...
    Text:
      title: Text
      description: Used to rappresent a text in a message or comment
      type: string
      pattern: '^[^\/\\]{0,100}$'
      example: "Hi!"
      minLength: 1
      maxLength: 100
...
    MsgText:
      title: Message Text
      description: rappresent a message text
      type: object
      properties:
        text: { $ref: "#/components/schemas/Text" }
        msgid: { $ref: "#/components/schemas/Identification" }
        send_time: { $ref: "#/components/schemas/Timestamp" }
      required:
        - text
        - msgid
        - send_time
    MsgPhoto:
      title: Message Photo
      description: rappresent a photo sended like a message
      type: object
      properties:
        photo: { $ref: "#/components/schemas/Photo" }
        caption: { $ref: "#/components/schemas/Text" }
        msgid: { $ref: "#/components/schemas/Identification" }
        send_time: { $ref: "#/components/schemas/Timestamp" }
      required:
        - photo
        - msgid
        - send_time
    BodyMessage:
      oneOf:
        - $ref: "#/components/schemas/MsgPhoto"
        - $ref: "#/components/schemas/MsgText"
    SentMessage:
      title: Sent message
      description: message posted by a user to another user ora a group
      type: object
      properties:
        body: { $ref: "#/components/schemas/BodyMessage" }
        status:
          description: Define the status of a message
          type: string
          enum: [not_arrive, one_check, two_check]
          example: one_check
    RcvMessage:
      title: Recived message
      description: message recived by a user or a group
      type: object
      properties:
        body: { $ref: "#/components/schemas/BodyMessage" }
        # The name of the user who sent the message
        sender_name: { $ref: "#/components/schemas/Name" }
    Message:
      oneOf:
        - $ref: "#/components/schemas/SentMessage"
        - $ref: "#/components/schemas/RcvMessage"
...
    Error:
      description: Object describe an error with the code and the message
      type: object
      properties:
        errors:
          description: Array of error codes generated by the operation
          type: array
          items:
            type: object
            description: Description of the error
            properties:
              code:
                description: Code error
                type: integer
                enum: [400, 401, 403, 404, 500]
                example: 404
              reason:
                type: string
                description: Explanation of the reason for the error which can be shown to a client user
                pattern: "^.*?$"
                example: Page not found
                minLength: 1
                maxLength: 100
              message:
                type: string
                description: More details of the error
                pattern: "^.*?$"
                example: "The requested URL was not found"
                minLength: 1
                maxLength: 100
            required:
              - code
              - reason
          minItems: 1
          maxItems: 100
      required:
        - errors
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The user is not allowed to perform the action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }